Moved getOrdersByUser method to UserService
