+++ src/main/java/com/orzechazo/eshop/services/UserService.java
@@ -1,6 +1,7 @@
 package com.orzechazo.eshop.services;
 
 import com.orzechazo.eshop.domain.Order;
+import com.orzechazo.eshop.domain.dto.OrderDto;
 import com.orzechazo.eshop.domain.dto.UserDto;
 
 import java.util.List;
@@ -8,6 +9,7 @@ import java.util.List;
 public interface UserService {
     List<UserDto> getAllUsers();
     UserDto getUserDtoByLogin(String login);
+    List<OrderDto> getOrdersByUser(String userLogin);
     UserDto createUser(UserDto userDto);
     UserDto updateUser(UserDto userDto);
     Order addOrder(String userLogin, Order order);
