+++ src/test/java/com/orzechazo/eshop/controllers/UserControllerTest.java
@@ -2,6 +2,7 @@ package com.orzechazo.eshop.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
+import com.orzechazo.eshop.domain.dto.OrderDto;
 import com.orzechazo.eshop.domain.dto.UserDto;
 import com.orzechazo.eshop.services.UserService;
 import org.junit.jupiter.api.BeforeEach;
@@ -59,6 +60,16 @@ class UserControllerTest {
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.login",equalTo("test")));
+    }
+    @Test
+    void getOrdersByUser() throws Exception {
+        List<OrderDto> orderDtos = List.of(OrderDto.builder().build(), OrderDto.builder().build());
+        when(userService.getOrdersByUser(any())).thenReturn(orderDtos);
+
+        mockMvc.perform(get("/users/testLogin/orders")
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$",hasSize(2)));
     }
 
     @Test
