+++ src/test/java/com/orzechazo/eshop/services/UserServiceImplTest.java
@@ -2,6 +2,7 @@ package com.orzechazo.eshop.services;
 
 import com.orzechazo.eshop.domain.Order;
 import com.orzechazo.eshop.domain.User;
+import com.orzechazo.eshop.domain.dto.OrderDto;
 import com.orzechazo.eshop.domain.dto.UserDto;
 import com.orzechazo.eshop.exceptions.BadRequestException;
 import com.orzechazo.eshop.exceptions.ResourceNotFoundException;
@@ -62,6 +63,26 @@ class UserServiceImplTest {
     void getUserByLoginNotFound() {
         Exception exception = assertThrows(ResourceNotFoundException.class,() -> userService.getUserDtoByLogin("test"));
         assertEquals("User: test doesn't exist in database.",exception.getMessage());
+    }
+    @Test
+    void getOrdersByUser() {
+        //given
+        Order order1 = new Order();
+        order1.setOrderId("order1");
+        Order order2 = new Order();
+        order2.setOrderId("order2");
+        User user = new User();
+        user.setLogin("login1");
+        user.setOrders(List.of(order1,order2));
+        List<String> orderIds = user.getOrders().stream().map(Order::getOrderId).toList();
+        when(userRepository.findByLogin(anyString())).thenReturn(Optional.of(user));
+        //when
+        List<OrderDto> returnedDtos = userService.getOrdersByUser("login1");
+        List<String> returnedDtosOrderIds = returnedDtos.stream().map(OrderDto::getOrderId).toList();
+        //then
+        assertEquals(2,returnedDtos.size());
+        assertThat(returnedDtosOrderIds).containsExactlyInAnyOrderElementsOf(orderIds);
+
     }
 
     @Test
